third_party/WebKit/LayoutTests/editing/deleting/delete-all-text-in-text-field-assertion.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<input id="sample" value="foo bar">
<script>
    test(() => {
        var sample = document.getElementById('sample');
        sample.focus();
        document.execCommand('selectAll');
        document.execCommand('delete');
        assert_equals(sample.value, '');
    }, 'document.execCommand works with INPUT');
</script>
